Add a Bitmap to record the tuple id range of all visible tuples
This is intended to add a Bitmap to record the range of AOTupleId
per segfile for all visible tuples. The range number is computed
as AOTupleId.rowNum / APPENDONLY_VISIMAP_MAX_RANGE indicating the
AOTupleId falls into this range, so it could return "visible"
directly instead of scanning the visimap table.